requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0

# Optimal Transport
POT>=0.9.0

# Medical imaging I/O
nibabel>=5.0.0
SimpleITK>=2.3.0
pydicom>=2.3.0

# Foundation models
timm>=0.9.0

# Evaluation metrics
scikit-learn>=1.3.0
scikit-image>=0.21.0
surface-distance>=0.1.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0

